[refactor] 프론트엔드 모듈화 및 리팩토링 - useSlide를 통한 custom hook 구현 - gql 분리 및 styled components 하단으로 위치 변경 - src/utils/model에 interface 정의 - 타입 지정
