+++ frontend/src/pages/SelectWorld/index.tsx
@@ -1,13 +1,29 @@
-/* eslint-disable jsx-a11y/click-events-have-key-events */
 /* eslint-disable jsx-a11y/alt-text */
-/* eslint-disable jsx-a11y/control-has-associated-label */
-/* eslint-disable react/button-has-type */
-import React, { useState, useEffect } from 'react';
-import { useQuery, gql } from '@apollo/client';
+import React, { useEffect } from 'react';
+import { useQuery } from '@apollo/client';
 import styled from 'styled-components';
-
+import { RouteComponentProps } from 'react-router';
 import { useRecoilState } from 'recoil';
+import { getWorldList } from '../../utils/query';
+import { WorldSelector } from '../../components/SelectWorld';
 import currentWorldState from '../../store/currentWorldState';
+
+const SelectWorld = (props: RouteComponentProps) => {
+    const { loading, error, data } = useQuery(getWorldList);
+    const [currentWorld, setCurrentWorld] = useRecoilState(currentWorldState);
+
+    if (loading) return <p>Loading...</p>;
+    if (error) return <p>Error :(</p>;
+    return (
+        <Backgroud>
+            <Logo>
+                <img src="/assets/logo.png" height="180px" />
+            </Logo>
+            <WorldSelector props={props} data={data.worldList} />
+        </Backgroud>
+    );
+};
+export default SelectWorld;
 
 const Logo = styled.div`
     position: absolute;
@@ -24,119 +40,3 @@ const Backgroud = styled.div`
     justify-content: center;
     align-items: center;
 `;
-
-const Selector = styled.div`
-    display: flex;
-    width: 600px;
-    flex-direction: row;
-    justify-content: space-between;
-    align-items: center;
-`;
-
-const SelectBtn = styled.div`
-    margin: 40px;
-    height: 60px;
-    width: 400px;
-    background-color: #c4c4c4;
-    text-align: center;
-    line-height: 60px;
-    border-radius: 20px;
-
-    font-family: Roboto;
-    font-style: normal;
-    font-weight: 150;
-    font-size: 40px;
-    font-weight: 500;
-`;
-
-const World = styled.div<{ thumbnail: string }>`
-    height: 400px;
-    width: 400px;
-    text-align: center;
-    line-height: 400px;
-    background-image: url('${(props) => props.thumbnail}.png');
-    background-size: 400px;
-    background-color: #c4c4c4;
-    border-radius: 20px;
-
-    font-family: Roboto;
-    font-style: normal;
-    font-weight: bold;
-    font-size: 80px;
-`;
-
-const getWorldList = gql`
-    query Query {
-        worldList {
-            id
-            uid
-            name
-            port
-            thumbnail
-        }
-    }
-`;
-
-interface IWorld {
-    id: number;
-    uid: number;
-    name: string;
-    port: number;
-    thumbnail: string;
-}
-
-const SelectWorld = ({ history }: any) => {
-    const { loading, error, data } = useQuery(getWorldList);
-    const [curWorld, setCurWorld] = useState<number>(0);
-    const [currentWorld, setCurrentWorld] =
-        useRecoilState<IWorld>(currentWorldState);
-
-    useEffect(() => {
-        if (currentWorld.name !== 'default')
-            history.push({ pathname: '/world' });
-    }, [currentWorld]);
-
-    if (loading) return <p>Loading...</p>;
-    if (error) return <p>Error :(</p>;
-
-    const nextClick = () => {
-        let cur = curWorld + 1;
-        if (cur >= data.worldList.length) cur = 0;
-        setCurWorld(cur);
-    };
-
-    const prevClick = () => {
-        let cur = curWorld - 1;
-        if (cur < 0) cur = data.worldList.length - 1;
-        setCurWorld(cur);
-    };
-
-    const redirectWorld = () => {
-        setCurrentWorld(data.worldList[curWorld]);
-    };
-
-    return (
-        <Backgroud>
-            <Logo>
-                <img src="/assets/logo.png" height="180px" />
-            </Logo>
-            <Selector>
-                <img
-                    src="/assets/prevbtn.png"
-                    onClick={prevClick}
-                    height="50px"
-                />
-                <World thumbnail={data.worldList[curWorld].thumbnail}>
-                    {data.worldList[curWorld].name}
-                </World>
-                <img
-                    src="/assets/nextbtn.png"
-                    onClick={nextClick}
-                    height="50px"
-                />
-            </Selector>
-            <SelectBtn onClick={redirectWorld}>선택</SelectBtn>
-        </Backgroud>
-    );
-};
-export default SelectWorld;
