+++ frontend/src/utils/model.ts
@@ -1,0 +1,7 @@
+export interface IWorld {
+    id: number;
+    uid: number;
+    name: string;
+    port: number;
+    thumbnail: string;
+}
