+++ frontend/src/utils/query.ts
@@ -1,0 +1,13 @@
+import { gql } from '@apollo/client';
+
+export const getWorldList = gql`
+    query Query {
+        worldList {
+            id
+            uid
+            name
+            port
+            thumbnail
+        }
+    }
+`;
